fix: rename test to tests
