+++ codemeta.json
@@ -66,7 +66,7 @@
             "givenName": "Robin"
         },
         {
-            "id": "",
+            "id": "https://orcid.org/0009-0007-9740-9070",
             "type": "Person",
             "affiliation": {
                 "type": "Organization",
