+++ pyproject.toml
@@ -11,9 +11,9 @@ classifiers = ["License :: OSI Approved :: GNU General Public License v3 or late
 requires-python = ">=3.10"
 dynamic = ["version", "description"]
 dependencies = [
-    "cmake~=3.31.6",
     "bsb[parallel,nest,neuron]~=6.0",
     "nestml~=8.0",
+    "cmake~=3.31.6",  # remove it next nestml release
     "odetoolbox==2.5.8",  # remove it next nestml release
     "matplotlib~=3.9",
     "elephant~=1.1"
