+++ tests/test_plots.py
