+++ tests/test_spiking_results.py
